Reorganize files, check extensions with regexps
