+++ lib/deppack/deppack.js
@@ -1,0 +1,107 @@
+'use strict';
+const sysPath = require('path');
+const fs = require('fs');
+const anymatch = require('anymatch');
+const modules = require('./modules');
+const promisify = require('../helpers').promisify;
+const mediator = require('../mediator');
+const xBrowserResolve = require('./resolve');
+const loadGlobalsAndStyles = require('./load-globs-styles');
+const explore = require('./explore');
+const processFiles = require('./process');
+const helpers = require('./helpers');
+const DepGraph = require('./dep-graph');
+const UsedShims = require('./used-shims');
+const ProjectModuleMap = require('./project-module-map');
+
+const readFile = promisify((path, callback) => fs.readFile(path, {encoding: 'utf8'}, callback));
+const dir = sysPath.dirname(sysPath.dirname(require.resolve('..')));
+
+const brunchRe = /brunch/;
+
+const packageRe = /node_modules/;
+const fileReflection = npmConfig => {
+  const isPackage = path => packageRe.test(path) && (npmConfig.npm.static || []).indexOf(path) === -1;
+  const isVendor = path => isPackage(path) ? false : anymatch(npmConfig.conventions.vendor, path);
+  const isApp = path => !anymatch(npmConfig.conventions.vendor, path);
+
+  return {isPackage, isVendor, isApp};
+};
+
+class Deppack {
+  constructor(config, json) {
+    if (!json.overrides) json.overrides = {};
+    if (!json.dependencies) json.dependencies = {};
+
+    const npmConfig = this.npmConfig = {
+      rootPackage: json,
+      conventions: config.conventions,
+      paths: config.paths,
+      npm: config.npm || {},
+      isProduction: mediator.isProduction,
+
+      globalPseudofile: '___globals___',
+
+      rootMainCache: {},
+
+      usedShims: new UsedShims(),
+      depGraph: new DepGraph(),
+      modMap: new ProjectModuleMap(config.modules.nameCleaner)
+    };
+
+    this.fileReflection = fileReflection(npmConfig);
+    this.resolve = xBrowserResolve.resolve(npmConfig.rootPackage, npmConfig.modMap, npmConfig.npm.aliases, npmConfig.globalPseudofile);
+
+    this.getPackageJson = filePath => {
+      return helpers.packageJson(filePath, npmConfig.rootPackage);
+    };
+  }
+
+  init() {
+    const npmConfig = this.npmConfig;
+    return xBrowserResolve.buildModMap(npmConfig.modMap, npmConfig.paths.watched, npmConfig.rootPackage)
+      .then(() => loadGlobalsAndStyles(npmConfig.paths.root, npmConfig.depGraph, npmConfig.globalPseudofile, npmConfig.npm, this.resolve));
+  }
+
+  exploreDeps(sourceFile) {
+    const npmConfig = this.npmConfig;
+    return explore(npmConfig.depGraph, npmConfig.modMap, npmConfig.usedShims, this.getPackageJson, npmConfig.isProduction, this.fileReflection, this.resolve)(sourceFile);
+  }
+
+  processFiles(root, files, processor) {
+    const npmConfig = this.npmConfig;
+    const requiredAliases = modules.requiredAliases(npmConfig.rootMainCache, this.getPackageJson, npmConfig.usedShims);
+    processFiles(npmConfig.npm.aliases || {}, npmConfig.globals || {}, npmConfig.usedShims, requiredAliases)(root, files, processor);
+  }
+
+  wrapInModule(filePath) {
+    filePath = sysPath.resolve('.', filePath);
+    const npmConfig = this.npmConfig;
+    return readFile(filePath).then(src => modules.wrap(this.getPackageJson, npmConfig.rootMainCache, filePath, src));
+  }
+
+  needsProcessing(file) {
+    return file.path.indexOf('node_modules') !== -1 &&
+      /\.(json|js)$/.test(file.path) &&
+      !brunchRe.test(file.path) &&
+      this.fileReflection.isPackage(file.path) ||
+      this.isNpm(file.path);
+  }
+
+  isNpm(path) {
+    return path.indexOf('node_modules') >= 0 &&
+      !brunchRe.test(path) && !/\.css$/.test(path) &&
+      this.fileReflection.isPackage(path) ||
+      path.indexOf(dir) === 0;
+  }
+
+  isNpmJSON(path) {
+    return this.isNpm(path) && /\.json$/.test(path);
+  }
+
+  isShim(path) {
+    return path.indexOf(dir) === 0;
+  }
+}
+
+module.exports = Deppack;
