+++ lib/deppack/explore.js
@@ -10,6 +10,8 @@ const shims = require('./shims');
 const helpers = require('./helpers');
 
 const not = helpers.not;
+const isRelative = helpers.isRelative;
+const makeRelative = helpers.makeRelative;
 
 const isJs = (fileList, path) => fileList.find(path).type === 'javascript';
 
@@ -21,7 +23,7 @@ const shouldIncludeDep = (getPackageJson, path) => {
     const root = dep.split('/')[0];
     const browser = typeof pkg.browser === 'object' ? pkg.browser : typeof pkg.browserify === 'object' ? pkg.browserify : {};
     // ignore optional dependencies
-    return (helpers.isRelative(dep) ||
+    return (isRelative(dep) ||
             root in deps ||
             browser[root] && browser[root] in deps ||
             root === currRoot ||
@@ -62,7 +64,7 @@ const depsFor = (fileReflection, getPackageJson) => {
   return (path, source) => {
     const allDeps = isVendor(path) ? [] : detective(source).filter(x => x && x.length > 0);
     const usesProcess = isVendor(path) ? false : shims.shouldIncludeProcess(source);
-    const deps = isApp(path) ? allDeps.filter(x => !helpers.isRelative(x)) : isPackage(path) ? allDeps.filter(shouldIncludeDep(getPackageJson, path)) : allDeps;
+    const deps = isApp(path) ? allDeps.filter(x => !isRelative(x)) : isPackage(path) ? allDeps.filter(shouldIncludeDep(getPackageJson, path)) : allDeps;
     if (usesProcess) deps.push('process');
     return deps;
   };
@@ -75,7 +77,7 @@ const exploreDeps = (depGraph, modMap, usedShims, getPackageJson, isProduction, 
   }
 
   const getDeps = depsFor(fileReflection, getPackageJson);
-  const shouldNotProcess = path => !isJs(fileList, path) || path.indexOf('.json') !== -1;
+  const shouldNotProcess = path => !isJs(fileList, path) || /\.json$/.test(path);
 
   const transform = sourceFile => {
     const source = envify(sourceFile.compiled, {NODE_ENV: isProduction ? 'production' : 'development'});
@@ -87,20 +89,21 @@ const exploreDeps = (depGraph, modMap, usedShims, getPackageJson, isProduction, 
     if (!sourceFile) return;
     const path = sourceFile.path;
 
+    modMap.add(path);
+
     if (shouldNotProcess(path)) return Promise.resolve(sourceFile);
 
     sourceFile = transform(sourceFile);
 
     const deps = getDeps(path, sourceFile.compiled);
 
-    modMap.add(path);
     if (depGraph.fileModsNotChanged(path, deps)) return Promise.resolve(sourceFile);
     depGraph.addFileMods(path, deps);
 
     const resPath = sysPath.resolve(path);
     const res = (i, cb) => browserResolve(i, {filename: resPath}, cb);
     return each(deps, res).then(fullPaths => {
-      const relPaths = fullPaths.filter(not(shims.isSimpleShim)).map(helpers.makeRelative);
+      const relPaths = fullPaths.filter(not(shims.isSimpleShim)).map(makeRelative);
       const fileShims = fullPaths.filter(shims.isShim).map(shims.actualShimName);
 
       const files = depGraph.getFilesToUnlink(path, depGraph.getFilesToCheckUnlink(path, relPaths));
