+++ lib/deppack/modules.js
@@ -150,7 +150,8 @@ const generateModule = (getPackageJson, rootMainCache, filePath, source) => {
 };
 
 const wrap = (getPackageJson, rootMainCache, filePath, source) => {
-  return cacheMain(getPackageJson, rootMainCache, filePath).then(() => generateModule(getPackageJson, rootMainCache, filePath, source));
+  return cacheMain(getPackageJson, rootMainCache, filePath)
+    .then(() => generateModule(getPackageJson, rootMainCache, filePath, source));
 };
