+++ lib/deppack/no-deppack.js
@@ -1,0 +1,37 @@
+'use strict';
+
+class NoDeppack {
+  init() {
+    return Promise.resolve({ components: [] });
+  }
+
+  exploreDeps(sourceFile) {
+    return Promise.resolve(sourceFile);
+  }
+
+  processFiles() {
+    return;
+  }
+
+  wrapInModule() {
+    return Promise.resolve();
+  }
+
+  needsProcessing() {
+    return false;
+  }
+
+  isNpm() {
+    return false;
+  }
+
+  isNpmJSON() {
+    return false;
+  }
+
+  isShim() {
+    return false;
+  }
+}
+
+module.exports = NoDeppack;
