+++ lib/deppack/process.js
@@ -4,7 +4,6 @@ const shims = require('./shims');
 const moduleDefinitions = require('./module-definitions');
 
 const word = /^[_\w]+$/;
-
 const glob = (globName, moduleName) => {
   const suffix = word.test(globName) ? `.${globName}` : `["${globName}"]`;
   return `window${suffix} = require("${moduleName}");`;
